ref(rcs): move some spot-tv only methods into spot-tvs rcs
